pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Alexey Serikov <serikov.alexey42@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.2"
sqlalchemy = "^2.0.25"
uvicorn = "^0.27.0.post1"
py-xid = "^1.0.0"
pytest = "^8.0.0"
httpx = "^0.26.0"
isort = "^5.13.2"
black = "^24.1.1"
flake8 = "^7.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
